backend/requirements.txt
fastapi
fastapi[standard]
requests
python-dotenv
groq
pinecone
llama-index
llama-index-core 
llama-index-llms-groq
llama-index-readers-docling
llama-index-node-parser-docling 
llama-index-embeddings-huggingface
llama-index-llms-huggingface-api
llama-index-readers-file
